fix(schema): constrain faq order to integer

+++ app/schemas/faq.ts
@@ -2,5 +2,5 @@ import { z } from 'zod/v4'
 import { createBaseWithSeoSchema } from './common'
 
 export const faqSchema = createBaseWithSeoSchema().extend({
-  order: z.number().default(999),
+  order: z.number().int().default(999),
 })
